Adding in a IEngineFacade to allow mocking and verification in consumer tests.
