+++ GoogleMapsApi/GoogleMapsApi.csproj
@@ -45,6 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="EngineFacade.cs" />
     <Compile Include="Engine\MapsAPIGenericEngine.cs" />
     <Compile Include="Engine\UnixTimeConverter.cs" />
     <Compile Include="Entities\Common\ILocationString.cs" />
@@ -120,6 +121,7 @@
     <Compile Include="Entities\TimeZone\Response\Status.cs" />
     <Compile Include="Entities\TimeZone\Response\TimeZoneResponse.cs" />
     <Compile Include="GoogleMaps.cs" />
+    <Compile Include="IEngineFacade.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Entities\Common\AddressLocation.cs" />
     <Compile Include="QueryStringParametersList.cs" />
